Move AutoComplete components to Beta module (#678)
Part of https://github.com/primer/view_components/issues/676

Renames:
 
1. `Primer::AutoComplete` to `Primer::Beta::AutoComplete`
2. `Primer::AutoComplete::Item` to `Primer::Beta::AutoComplete::Item`

I had to update a bunch of configuration files since having a `stories/primer/beta/auto_complete` directory was making zeitwerk assume that `Primer::Beta::AutoComplete` was a module.
The changes are:

1. Stop loading stories files when generating docs (prevents the zeitwerk issue)
2. Update `docs-build` CI to set an env var that will do the above
3. Creating `stories/primer/beta/auto_complete.rb` to explicity set `Primer::Beta::AutoComplete` as a class for the stories build
