+++ .github/workflows/docs-build.yml
@@ -34,6 +34,8 @@ jobs:
           bundle exec rake utilities:build
           bundle exec rake docs:build
           bundle exec rake statuses:dump
+        env:
+          SKIP_STORYBOOK_PRELOAD: 1
       - name: Commit & Push Docs Data
         run: |
           git config --local user.email "actions@github.com"
