+++ CHANGELOG.md
@@ -36,6 +36,12 @@ The category for changes related to documentation, testing and tooling. Also, fo
 
     *Kate Higa*
 
+* Renames:
+  * `Primer::AutoComplete` to `Primer::Beta::AutoComplete`
+  * `Primer::AutoComplete::Item` to `Primer::Beta::AutoComplete::Item`
+
+    *Manuel Puyol*
+
 ### Misc
 
 * Update `doc_examples_axe_test` to exclude non-standalone components and fix `Markdown` example.
