+++ static/statuses.json
@@ -1,12 +1,12 @@
 {
   "Primer::Alpha::ButtonMarketing": "alpha",
-  "Primer::AutoComplete": "beta",
-  "Primer::AutoComplete::Input": "alpha",
-  "Primer::AutoComplete::Item": "beta",
   "Primer::AvatarComponent": "beta",
   "Primer::AvatarStackComponent": "beta",
   "Primer::BaseButton": "beta",
   "Primer::BaseComponent": "beta",
+  "Primer::Beta::AutoComplete": "beta",
+  "Primer::Beta::AutoComplete::Input": "alpha",
+  "Primer::Beta::AutoComplete::Item": "beta",
   "Primer::Beta::Text": "beta",
   "Primer::BlankslateComponent": "beta",
   "Primer::BorderBoxComponent": "beta",
