+++ stories/primer/beta/auto_complete.rb
@@ -1,0 +1,4 @@
+# frozen_string_literal: true
+
+# needed for zeitwerk to not assume AutoComplete is a module instead of class.
+class Primer::Beta::AutoComplete < Primer::Component; end
